Non-records should return empty lists for 'keys'.

+++ src/libawkward/array/UnionArray.cpp
@@ -383,7 +383,7 @@ namespace awkward {
           }
         }
         if (!found) {
-          out.erase(out.begin() + j);
+          out.erase(out.begin() + (size_t)j);
         }
       }
     }
